refactor: y/n 검증 로직 뷰 패키지로 이동
