+++ src/main/java/view/InputView.java
@@ -17,8 +17,9 @@ public class InputView {
             .collect(Collectors.toList());
     }
 
-    public String askHitCommand(final String playerName) {
+    public boolean askHitCommand(final String playerName) {
         System.out.println(playerName + "는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)");
-        return scanner.nextLine();
+        String command = scanner.nextLine();
+        return HitCommand.from(command);
     }
 }
